+++ libfreerdp/utils/svc_plugin.c
@@ -395,7 +395,7 @@ int svc_plugin_send(rdpSvcPlugin* plugin, STREAM* data_out)
 {
 	uint32 error = 0;
 
-	DEBUG_SVC("length %d", stream_get_length(data_out));
+	DEBUG_SVC("length %d", (int) stream_get_length(data_out));
 
 	error = plugin->channel_entry_points.pVirtualChannelWrite(plugin->priv->open_handle,
 		stream_get_data(data_out), stream_get_length(data_out), data_out);
